+++ README.md
@@ -1,1 +1,3 @@
 # daibouken
+
+Successor to [chousen](https://github.com/RawToast/chousen-server)
