+++ src/test/scala/daibouken/ExploratoryTest.scala
@@ -1,0 +1,7 @@
+package daibouken
+
+import org.scalatest.{Matchers, WordSpec}
+
+class ExploratoryTest extends WordSpec with Matchers {
+
+}
